+++ TestPinb/OOP/Light.h
@@ -16,9 +16,9 @@ http://pinballhomemade.blogspot.com.br
 class Light : public Output
 {
 public:
-	Light(const char *szName, PinballMaster *pinball, int portNumber);
+	Light(const char *szName, PinballMaster *pinball, uint8_t portNumber);
 	virtual ~Light();
-	virtual bool NotifyEvent(PinballObject *sender, int event, int valueToSend);
+	virtual bool NotifyEvent(PinballObject *sender, uint8_t event, uint8_t valueToSend);
 
 	void StartBlink(long value=700);
 	void EndBlink();
